fix: Add explicit Float conversion for LinearProgressIndicator
- Convert currentStep to Float explicitly with .toFloat()
- Fixes type inference issue in Kotlin
- Progress now correctly calculates as Float value

+++ app/src/main/java/com/novahypnose/ancrage/ui/screens/creation/CreateAnchorScreen.kt
@@ -58,7 +58,7 @@ fun CreateAnchorScreen(
         ) {
             // Indicateur de progression
             LinearProgressIndicator(
-                progress = uiState.currentStep / 4f,
+                progress = uiState.currentStep.toFloat() / 4f,
                 modifier = Modifier.fillMaxWidth()
             )
